[data] Keep row slicing coverage with map operator tests
